components/appkit-commons/pyproject.toml
[project]
name = "appkit-commons"
version = "0.5.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Jens Rehpöhler" }]
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "cryptography>=46.0.2",
    "pydantic-settings>=2.10.1",
    "pyyaml==6.0.2",
    "sqlalchemy-utils==0.42.0",
    "sqlalchemy==2.0.41",
]

[project.optional-dependencies]
azure = [
    "azure-identity==1.23.1",
    "azure-keyvault-secrets==4.10.0",
]


[tool.setuptools.packages.find]
where = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/appkit_commons"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
